requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
jinja2==3.1.3
python-dotenv==1.0.1
requests==2.31.0
